created new component LoginPageForm.tsx and refactored LoginPage.tsx
