+++ src/AdminPanel/modules/LoginPage/LoginPageForm/index.ts
@@ -1,0 +1,1 @@
+export * from './LoginPageForm';
